move toggle to popover
